+++ test/mpi/ReinitialiserTests/MultiStageStabilisedReinitTest.jl
@@ -36,11 +36,11 @@ function main(distribute,mesh_partition)
     )
   end
 
-  ls_evo = CutFEMEvolver(V_φ,Ωs,dΩ,h)
-  reinit1 = StabilisedReinitialiser(V_φ,Ωs,dΩ,h;
+  ls_evo = CutFEMEvolver(V_φ,dΩ,h)
+  reinit1 = StabilisedReinitialiser(V_φ,dΩ,h;
     stabilisation_method=ArtificialViscosity(1.5h),
     nls = GridapSolvers.NewtonSolver(LUSolver();maxiter=50,rtol=1.e-14,verbose=i_am_main(ranks)))
-  reinit2 = StabilisedReinitialiser(V_φ,Ωs,dΩ,h;
+  reinit2 = StabilisedReinitialiser(V_φ,dΩ,h;
     stabilisation_method=InteriorPenalty(V_φ),
     nls = GridapSolvers.NewtonSolver(LUSolver();maxiter=50,rtol=1.e-14,verbose=i_am_main(ranks)))
   ls_reinit = GridapTopOpt.MultiStageStabilisedReinitialiser([reinit1,reinit2])